novo

+++ tab.py
@@ -1,3 +1,3 @@
-n = float(input('digite um número:'))
+n = int(input('digite um número:'))
 for i in range(1,11):
      print(f'{n} X {i} = {n*i}')
